+++ apps/fluent-tester/src/E2E/ButtonExperimental/specs/ButtonExperimental.spec.win.ts
@@ -1,6 +1,6 @@
-import NavigateAppPage from '../../common/NavigateAppPage.win';
+import NavigateAppPage from '../../common/NavigateAppPage';
 import ButtonExperimentalPageObject, { ButtonSelector } from '../pages/ButtonExperimentalPageObject';
-import { ComponentSelector } from '../../common/BasePage.win';
+import { ComponentSelector, Platform } from '../../common/BasePage';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT, BUTTON_A11Y_ROLE, Keys } from '../../common/consts';
 import { BUTTON_ACCESSIBILITY_LABEL, BUTTON_TEST_COMPONENT_LABEL } from '../../../TestComponents/Button/consts';
 
@@ -13,7 +13,7 @@ describe('Experimental Button Testing Initialization', function () {
 
   it('Click and navigate to Button test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    await ButtonExperimentalPageObject.scrollToComponentButton();
+    await ButtonExperimentalPageObject.scrollToComponentButton(Platform.Win32);
     await ButtonExperimentalPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */