+++ apps/fluent-tester/src/E2E/FocusTrapZone/specs/FocusTrapZone.spec.win.ts
@@ -1,6 +1,7 @@
-import NavigateAppPage from '../../common/NavigateAppPage.win';
+import NavigateAppPage from '../../common/NavigateAppPage';
 import FocusTrapZonePageObject from '../pages/FocusTrapZonePageObject.win';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
+import { Platform } from '../../common/BasePage';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('FocusTrapZone Testing Initialization', function () {
@@ -11,7 +12,7 @@ describe('FocusTrapZone Testing Initialization', function () {
 
   it('Click and navigate to FocusTrapZone test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    await FocusTrapZonePageObject.scrollToComponentButton();
+    await FocusTrapZonePageObject.scrollToComponentButton(Platform.Win32);
     await FocusTrapZonePageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */