+++ apps/fluent-tester/src/E2E/FocusZone/specs/FocusZone.spec.win.ts
@@ -1,6 +1,7 @@
-import NavigateAppPage from '../../common/NavigateAppPage.win';
-import FocusZonePageObject from '../pages/FocusZonePageObject.win';
+import NavigateAppPage from '../../common/NavigateAppPage';
+import FocusZonePageObject from '../pages/FocusZonePageObject';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
+import { Platform } from '../../common/BasePage';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('FocusZone Testing Initialization', function () {
@@ -11,7 +12,7 @@ describe('FocusZone Testing Initialization', function () {
 
   it('Click and navigate to FocusZone test page', async () => {
     /* Scroll to component test page button in scrollview if not already visible*/
-    await FocusZonePageObject.scrollToComponentButton();
+    await FocusZonePageObject.scrollToComponentButton(Platform.Win32);
     await FocusZonePageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */