+++ apps/fluent-tester/src/E2E/Persona/specs/Persona.spec.macos.ts
@@ -1,5 +1,5 @@
-import NavigateAppPage from '../../common/NavigateAppPage.macos';
-import PersonaPageObject from '../pages/PersonaPageObject.macos';
+import NavigateAppPage from '../../common/NavigateAppPage';
+import PersonaPageObject from '../pages/PersonaPageObject';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
@@ -10,8 +10,6 @@ describe('Persona Testing Initialization', function () {
   });
 
   it('Click and navigate to Persona test page', async () => {
-    /* Scroll to component test page button in scrollview if not already visible*/
-    await PersonaPageObject.scrollToComponentButton();
     await PersonaPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */