+++ apps/fluent-tester/src/E2E/Tabs/specs/Tabs.spec.macos.ts
@@ -1,5 +1,5 @@
-import NavigateAppPage from '../../common/NavigateAppPage.macos';
-import TabsPageObject from '../pages/TabsPageObject.macos';
+import NavigateAppPage from '../../common/NavigateAppPage';
+import TabsPageObject from '../pages/TabsPageObject';
 import { BOOT_APP_TIMEOUT, PAGE_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
@@ -10,8 +10,6 @@ describe('Tabs Testing Initialization', function () {
   });
 
   it('Click and navigate to Tabs test page', async () => {
-    /* Scroll to component test page button in scrollview if not already visible*/
-    await TabsPageObject.scrollToComponentButton();
     await TabsPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */