+++ apps/fluent-tester/src/E2E/TabsExperimental/specs/TabsExperimental.spec.macos.ts
@@ -1,5 +1,5 @@
-import NavigateAppPage from '../../common/NavigateAppPage.macos';
-import ExperimentalTabsPageObject from '../pages/ExperimentalTabsPageObject.macos';
+import NavigateAppPage from '../../common/NavigateAppPage';
+import ExperimentalTabsPageObject from '../pages/ExperimentalTabsPageObject';
 import { BOOT_APP_TIMEOUT, PAGE_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
@@ -10,8 +10,6 @@ describe('Experimental Tabs Testing Initialization', function () {
   });
 
   it('Click and navigate to Experimental Tabs test page', async () => {
-    /* Scroll to component test page button in scrollview if not already visible*/
-    await ExperimentalTabsPageObject.scrollToComponentButton();
     await ExperimentalTabsPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */