+++ packages/experimental/Dropdown/src/Option/Option.types.ts
@@ -1,15 +1,31 @@
 import type { IViewProps } from '@fluentui-react-native/adapters';
+import { FontTokens, IBorderTokens, IColorTokens, LayoutTokens } from '@fluentui-react-native/framework';
 import { IPressableHooks, IWithPressableOptions } from '@fluentui-react-native/interactive-hooks';
 import { TextProps } from '@fluentui-react-native/text';
+import { ColorValue } from 'react-native';
 import { SvgProps } from 'react-native-svg';
 
 export const optionName = 'Option';
 
-export interface OptionTokens {}
+export interface OptionTokens extends FontTokens, IBorderTokens, IColorTokens, LayoutTokens {
+  /**
+   * Color of the checkmark icon on the Option indicating Option is selected
+   */
+  checkmarkColor?: ColorValue;
 
-export interface OptionProps extends IWithPressableOptions<IViewProps> {}
+  /**
+   * Height and width of the checkmark icon on the Option indicating Option is selected
+   */
+  checkmarkSize?: number;
 
-export interface OptionState extends IPressableHooks<OptionProps & React.ComponentPropsWithRef<any>> {}
+  /**
+   * Spacing, in pixels, between the label and icons
+   */
+  spacingContentIcon?: number;
+}
+
+export type OptionProps = IWithPressableOptions<IViewProps>;
+export type OptionState = IPressableHooks<OptionProps & React.ComponentPropsWithRef<any>>;
 
 export interface OptionSlotProps {
   root: IViewProps;